requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
scikit-learn==1.3.2
joblib==1.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==1.12.1
python-dotenv==1.0.0
pandas==2.1.3
pydantic==2.5.0
requests==2.31.0
